Add to cart functionality added and links corrected on index page
Items are added to the cart and can be removed. Quantity updates on the card page.  I need to fix the price not displaying correctly.

Added a hover over cart icon effect so you can see the items in your cart and when you move away from that window is disappears

Links on index.html page redirected to  the specific category on the lookbook page